% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_elegans.R
\name{add_elegans}
\alias{add_elegans}
\title{add_elegans function}
\usage{
add_elegans(input_list)
}
\arguments{
\item{input_list}{A list of strings which represent species names}
}
\value{
A vector of species names to which the species C. Elegans has been added
}
\description{
Add C. Elegans species name into a vector of string
}
\details{
Belong to the server function in the show button section
}
\examples{
#...
}
