% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seq_id.R
\name{add_seq_id}
\alias{add_seq_id}
\title{add_seq_id function}
\usage{
add_seq_id(df, species)
}
\arguments{
\item{df}{A species specific data frame}

\item{species}{A string that represents a species name}
}
\value{
A data frame with seq_id column added
}
\description{
Add seq_id column to top fragments tables
}
\details{
Belong to the pipeline function in step 11 : create aligned sequences, aligned genes and links tables for gggenomes
}
\examples{
#...
}
