% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_colors.R
\name{assign_colors}
\alias{assign_colors}
\title{assign_colors function}
\usage{
assign_colors(genes, prev_colors)
}
\arguments{
\item{genes}{A data frame of genes names}

\item{prev_colors}{A vector of colors}
}
\value{
A list containing a data frame and a map
}
\description{
Assign colors to orthogroups in a vector while plotting
}
\details{
Belong to the server function in every section
}
\examples{
#...
}
