% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_end.R
\name{calculate_end}
\alias{calculate_end}
\title{calculate_end function}
\usage{
calculate_end(df)
}
\arguments{
\item{df}{A data frame issued from halLiftover linux command}
}
\value{
A data frame with end column calculated
}
\description{
Calculate the end column of a PSL file
}
\details{
Belong to the sort_halLiftover_data function in the pipeline function of the step 1 : retrieve the homologous fragments of the C. elegans sequence of inquiry
}
\examples{
#...
}
