% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dfs.R
\name{combine_dfs}
\alias{combine_dfs}
\title{combine_dfs function}
\usage{
combine_dfs(individual_dfs)
}
\arguments{
\item{individual_dfs}{A list of data frames}
}
\value{
A data frame
}
\description{
Combine individual data frames
}
\details{
Belong to the show_fct function in the server function in the show button section
}
\examples{
#...
}
