% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control}
\alias{control}
\title{control function}
\usage{
control(genes_file, extras_genes)
}
\arguments{
\item{genes_file}{An aligned genes data frame}

\item{extras_genes}{An extras genes data frame}
}
\value{
An extras genes data frame
}
\description{
Remove same genes in extras fragments files
}
\details{
Belong to the pipeline function in step 10 : remove same genes from extras fragments and place them into top fragments
}
\examples{
#...
}
