% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_individual_dfs.R
\name{create_individual_dfs}
\alias{create_individual_dfs}
\title{create_individual_dfs function}
\usage{
create_individual_dfs(links)
}
\arguments{
\item{links}{A links data frame}
}
\value{
A list of data frames
}
\description{
Create individual data frames from links table
}
\details{
Belong to the show_fct in the server function in the show button section
}
\examples{
#...
}
