% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_region.R, R/export_fct.R,
%   R/pipeline_fct.R
\docType{data}
\name{folder_ubuntu}
\alias{folder_ubuntu}
\title{create_region function}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
folder_ubuntu

folder_ubuntu

folder_ubuntu
}
\arguments{
\item{chr}{A string as a chromosome name}

\item{start}{A integer as the start}

\item{end}{A integer as the end}

\item{gap}{A integer as the maximum length between two fragments}

\item{filtering_percentage}{A integer as the filtering percentage}

\item{gname}{A string as the gene name or NULL}

\item{sorted_data}{A data frame issued from HalLiftover linux command}

\item{species_name}{A species string name}

\item{region}{The inquiry region, a data frame generated with create_region function}
}
\value{
A single row table

Exported tables into the ubuntu folder

A list of data frames : sequences, aligned_genes, links and aligned_sequences
}
\description{
Create the region inquiry : seqnames, start, end, gap, filtering_value, gene_name

Export the sorted tables from halLiftover linux command in the ubuntu folder

Generate three types of tables to be used in gggenomes: aligned_sequences used in gggenomes, aligned_genes, and links along with an aligned sequences table to be display
}
\details{
Belong to the server function in search, zoom in and out button sections

Belong to the pipeline function in step 2 : find pairwise alignment between phylogenetically adjacent species using the homologous fragments

Belong to the server function in every section
}
\examples{

# Coordinates filling case
# region <- create_region("X", 4271165, 4274216, NULL, 200, 16)
# Gene selection case
# region <- create_region("X", 4271165, 4274216, "mec-17", 200, 16)
#...

# Generate the necessary three tables for gggenomes: aligned_genes, aligned_sequences and links
# data <- pipeline_fct(region)

}
\keyword{datasets}
