% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_fct.R
\name{genes_fct}
\alias{genes_fct}
\title{genes_fct function}
\usage{
genes_fct(species, final_species)
}
\arguments{
\item{species}{A species specific data frame}

\item{final_species}{A data frame of the top fragments}
}
\value{
An aligned genes data frame
}
\description{
Intersect genes from aligned fragments tables
}
\details{
Belong to the pipeline function in step 6 : find genes/piece of genes in the aligned regions
}
\examples{
#...
}
