% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_halLiftover_folder.R
\name{get_halLiftover_path}
\alias{get_halLiftover_path}
\title{Function to get the path to the 'halLiftover' command}
\usage{
get_halLiftover_path()
}
\value{
The halLiftover cmd path
}
\description{
Function to get the path to the 'halLiftover' command
}
\examples{
#...
}
