% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ubuntu_path.R
\name{get_ubuntu_path}
\alias{get_ubuntu_path}
\title{Function to get the path to the 'ubuntu' folder within the package}
\usage{
get_ubuntu_path()
}
\value{
The ubuntu folder path
}
\description{
Function to get the path to the 'ubuntu' folder within the package
}
\examples{
#...
}
