% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggg_plot.R
\name{ggg_plot}
\alias{ggg_plot}
\title{ggg_plot function}
\usage{
ggg_plot(seqs, genes, links, region, elegans_file)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}

\item{links}{A links data frame}

\item{region}{An inquiry region data frame}

\item{elegans_file}{The C. Elegans genes data frame}
}
\value{
A gggenomes plot
}
\description{
Create gggenomes plots in functions of the three main tables
}
\details{
Belong to the the server function in every section
}
\examples{
#...
}
