% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_links.R
\name{has_links}
\alias{has_links}
\title{has_links function}
\usage{
has_links(seqs, links)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{links}{A links data frame}
}
\value{
A boolean : TRUE if there is links and FALSE if there is not
}
\description{
Test if there is links
}
\details{
Belong to the ggg_plot function in the server function in every section
}
\examples{
#...
}
