% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_links.R
\name{make_links}
\alias{make_links}
\title{make_links function}
\usage{
make_links(list_df, new_selection)
}
\arguments{
\item{list_df}{A list of aligned fragments data frames}

\item{new_selection}{A list of species selected by the user}
}
\value{
A list
}
\description{
Make new links according to the user species selection
}
\details{
Belong to the server function in the show button section
}
\examples{
#...
}
