% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{micro}
\alias{micro}
\title{micro function}
\usage{
micro(genes, species)
}
\arguments{
\item{genes}{An aligned genes data frame}

\item{species}{A species specific data frame}
}
\value{
An aligned genes data frame with CDS gene type
}
\description{
Transform the aligned genes data frame data into microsynteny data
}
\details{
Belong to the server function in the microsynteny button section
}
\examples{
#...
}
