% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_5.R
\name{plot_5}
\alias{plot_5}
\title{plot_5 function}
\usage{
plot_5(seqs, genes, region)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}

\item{region}{A region inquiry data frame}
}
\value{
A gggenomes plot
}
\description{
Generate plot case 5 : no C. Elegans genes and no links
}
\details{
Belong to the ggg_plot function in the server function in every section
}
\examples{
#...
}
