% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslToDataFrame2.R
\name{pslToDataFrame2}
\alias{pslToDataFrame2}
\title{pslToDataFrame2 function}
\usage{
pslToDataFrame2(list_df, species_names)
}
\arguments{
\item{elegans_species}{A PSL format file}

\item{species_name}{A string that represents a species name}
}
\value{
A data frame of aligned fragments
}
\description{
From PSL file format to data frame format case 2 : species as queries
}
\details{
Belong to the pipeline function in step 2 : find pairwise alignment between phylogenetically adjacent species using the homologous fragments
}
\examples{
#...
}
