% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_explanation_in.R
\name{render_explanation_in}
\alias{render_explanation_in}
\title{render_explanation_in function}
\usage{
render_explanation_in(output, region)
}
\arguments{
\item{output}{An output from a shiny app session}

\item{region}{A region inquiry}
}
\value{
A string for the explanation panel
}
\description{
Explanations related to the zoom in button
}
\details{
Belong to the server function in zoom in section
}
\examples{
#...
}
