% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_explanation_out.R
\name{render_explanation_out}
\alias{render_explanation_out}
\title{render_explanation_out function}
\usage{
render_explanation_out(output, region)
}
\arguments{
\item{output}{An output from a shiny app session}

\item{region}{A region inquiry}
}
\value{
A string for the explanation panel
}
\description{
Explanations related to the zoom out button
}
\details{
Belong to the server function in zoom out section
}
\examples{
#...
}
