% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_explanations.R
\name{render_explanations}
\alias{render_explanations}
\title{render_explanations function}
\usage{
render_explanations(output, genes, seqs)
}
\arguments{
\item{output}{An output from a shny app session}

\item{genes}{An aligned genes data frame}

\item{seqs}{An aligned sequences data frame}
}
\value{
Display messages on the explanation panel
}
\description{
Render explanations on the UI
}
\details{
Belong to the server function in every section
}
\examples{
#...
}
