% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_blockSizes.R
\name{reverse_blockSizes}
\alias{reverse_blockSizes}
\title{reverse_blockSizes function}
\usage{
reverse_blockSizes(df)
}
\arguments{
\item{df}{A data frame of aligned fragments from halLiftover}
}
\value{
A data frame with reverse blockSizes
}
\description{
Reverse blocksizes depending of the strand
}
\details{
Belong to the sort_haliftover_data function in the pipeline function in step 1 : retrieve the homologous fragments of the C. Elegans sequence of inquiry
}
\examples{
#...
}
