% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs_change.R
\name{seqs_change}
\alias{seqs_change}
\title{seqs_change function}
\usage{
seqs_change(seqs, genes)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}
}
\value{
An aligned sequences data frame used in gggenomes
}
\description{
Change the aligned sequences data frame coordinates to be used in gggenomes according to extended genes
}
\details{
Belong to the server function in the full genes extension button section
}
\examples{
#...
}
