% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_haliftover_data.R
\name{sort_haliftover_data}
\alias{sort_haliftover_data}
\title{sort_haliftover_data function}
\usage{
sort_haliftover_data(df)
}
\arguments{
\item{df}{A data frame from halLiftover command}
}
\value{
A formatted data frame
}
\description{
Formatting the aligned fragments data frame from halLiftover command
}
\details{
Belong to the pipeline function in step 1 : retrieve the homologous fragments of the C. Elegans sequence of inquiry
}
\examples{
#...
}
