% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_start_end.R
\name{split_start_end}
\alias{split_start_end}
\title{split_start_end function}
\usage{
split_start_end(df)
}
\arguments{
\item{df}{A data frame issued from halLiftover linux command}
}
\value{
A data frame with unlist start and end columns
}
\description{
Unlist start and end columns
}
\details{
Belong to the sort_haliftover_data function in the pipeline function in step 1 : retrieve the homologous fragments of the C. Elegans sequence of inquiry
}
\examples{
#...
}
