% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{stitch function}
\usage{
stitch(raw_coord, gap_max)
}
\arguments{
\item{raw_coord}{An aligned raw fragments data frame}

\item{gap_max}{A gap value as integer}
}
\value{
An aligned assembled fragments data frame
}
\description{
Stitch adjacent fragments together below a certain gap
}
\details{
Belong to the pipeline function in step 3 : assemble adjacent fragments according to a gap value
}
\examples{
#...
}
