% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_intersect.R
\name{subset_by_intersect}
\alias{subset_by_intersect}
\title{subset_by_intersect function}
\usage{
subset_by_intersect(species, aligned_coord_species)
}
\arguments{
\item{species}{A species specific data frame}

\item{aligned_coord_species}{An aligned fragments data frame}
}
\value{
A data frame of intersected genes
}
\description{
Intersect the aligned fragments data frame
}
\details{
Belong to the genes_fct function in the pipeline function in step 6 : find genes/piece of genes in the aligned regions
}
\examples{
#...
}
