% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_overlap.R
\name{subset_by_overlap}
\alias{subset_by_overlap}
\title{subset_by_overlap function}
\usage{
subset_by_overlap(species, aligned_coord_species)
}
\arguments{
\item{species}{A species specific data frame}

\item{aligned_coord_species}{An aligned fragments data frame}
}
\value{
A data frame of overlapped genes
}
\description{
Overlap the aligned fragments data frame
}
\details{
Belong to the check_overlapping function in the pipeline function in step 8 : check if there is some "hidden genes" inside the intersected genes found
}
\examples{
#...
}
